[TWINFACES-288] feat: resolve PR comments
