+++ src/entities/factory/api/index.ts
@@ -1,2 +1,4 @@
 export * from "./types";
 export * from "./api-service";
+export * from "./use-fetch-factory-data";
+export * from "./use-update-factory";
