+++ src/entities/factory/libs/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks";
 export * from "./helpers";
+export * from "./factory-context";
