+++ src/screens/factories/views/index.ts
